refactor(urls): split server_router.rs by responsibility (#4310)
Decompose the 3466-line server_router.rs into responsibility-based
submodules under server_router/ (builder, compile, dispatch,
introspection, registration, router_impls, types, tests) following
Rust 2024 Edition module conventions. Public API preserved via pub use
re-exports.

Fixes #4310
